Alien ship
Added a smaller alien ship that is scaled down by exactly .5
Placed the large alien ship on the screen once the 2nd level is reached.
Placed the smaller alien ship on the screen on all levels above 2.
Alien ships will now be destroyed when they impact an object.
Alien ships now have sounds that are played until they are destroyed.